+++ contracts/nois-proxy/src/contract.rs
@@ -29,10 +29,17 @@ use crate::state::{Config, OperationalMode, ALLOWLIST, ALLOWLIST_MARKER, CONFIG,
 
 pub const REPLAY_ID_CALLBACK: u64 = 456;
 
+/// 10 years in seconds
+const TEN_YEARS_S: u64 = 10 * 3600 * 24 * 365;
+
+/// If not set otherwise, min_after is the genesis time of Nois mainnet
+const MIN_AFTER_FALLBACK: Timestamp = Timestamp::from_seconds(1680015600);
+const MAX_AFTER_FALLBACK: Timestamp = MIN_AFTER_FALLBACK.plus_seconds(TEN_YEARS_S);
+
 #[cfg_attr(not(feature = "library"), entry_point)]
 pub fn instantiate(
     deps: DepsMut,
-    _env: Env,
+    env: Env,
     _info: MessageInfo,
     msg: InstantiateMsg,
 ) -> StdResult<Response> {
@@ -63,6 +70,8 @@ pub fn instantiate(
         nois_beacon_price_updated: Timestamp::from_seconds(0),
         mode,
         allowlist_enabled: Some(allowlist_enabled),
+        min_after: Some(env.block.time),
+        max_after: Some(env.block.time.plus_seconds(TEN_YEARS_S)),
     };
 
     CONFIG.save(deps.storage, &config)?;
@@ -78,10 +87,20 @@ pub fn instantiate(
         .add_attribute("test_mode", test_mode.to_string()))
 }
 
-// This no-op migrate implementation allows us to upgrade within the 0.7 series.
-// No state changes expected.
 #[cfg_attr(not(feature = "library"), entry_point)]
-pub fn migrate(_deps: DepsMut, _env: Env, _msg: Empty) -> StdResult<Response> {
+pub fn migrate(deps: DepsMut, env: Env, _msg: Empty) -> StdResult<Response> {
+    let mut config = CONFIG.load(deps.storage)?;
+
+    // If unset, set min_after and max_after to now and now+10y
+    if config.min_after.is_none() {
+        config.min_after = Some(env.block.time);
+    }
+    if config.max_after.is_none() {
+        config.max_after = Some(env.block.time.plus_seconds(TEN_YEARS_S));
+    }
+
+    CONFIG.save(deps.storage, &config)?;
+
     Ok(Response::default().add_attribute(ATTR_ACTION, "migtrate"))
 }
 
@@ -104,6 +123,8 @@ pub fn execute(
             callback_gas_limit,
             mode,
             allowlist_enabled,
+            min_after,
+            max_after,
         } => execute_set_config(
             deps,
             info,
@@ -115,6 +136,8 @@ pub fn execute(
             callback_gas_limit,
             mode,
             allowlist_enabled,
+            min_after,
+            max_after,
         ),
         ExecuteMsg::GetRandomnessAfter { after, job_id } => {
             execute_get_randomness_after(deps, env, info, after, job_id)
@@ -164,6 +187,7 @@ fn execute_get_randomness_after(
     job_id: String,
 ) -> Result<Response, ContractError> {
     let config = CONFIG.load(deps.storage)?;
+
     execute_get_randomness_impl(
         deps,
         env,
@@ -191,6 +215,16 @@ pub fn execute_get_randomness_impl(
     let allowlist_enabled = config.allowlist_enabled.unwrap_or(false);
     if allowlist_enabled && !ALLOWLIST.has(deps.storage, &info.sender) {
         return Err(ContractError::SenderNotAllowed);
+    }
+
+    let min_after = config.min_after.unwrap_or(MIN_AFTER_FALLBACK);
+    if after < min_after {
+        return Err(ContractError::AfterTooLow { min_after, after });
+    }
+
+    let max_after = config.max_after.unwrap_or(MAX_AFTER_FALLBACK);
+    if after > max_after {
+        return Err(ContractError::AfterTooHigh { max_after, after });
     }
 
     let packet = InPacket::RequestBeacon {
@@ -255,6 +289,8 @@ fn execute_set_config(
     callback_gas_limit: Option<u64>,
     mode: Option<OperationalMode>,
     allowlist_enabled: Option<bool>,
+    min_after: Option<Timestamp>,
+    max_after: Option<Timestamp>,
 ) -> Result<Response, ContractError> {
     let config = CONFIG.load(deps.storage)?;
 
@@ -276,6 +312,8 @@ fn execute_set_config(
         callback_gas_limit,
         mode,
         allowlist_enabled,
+        min_after,
+        max_after,
     )
 }
 
@@ -351,6 +389,8 @@ pub fn sudo(deps: DepsMut, env: Env, msg: SudoMsg) -> Result<Response, ContractE
             callback_gas_limit,
             mode,
             allowlist_enabled,
+            min_after,
+            max_after,
         } => sudo_set_config(
             deps,
             env,
@@ -361,6 +401,8 @@ pub fn sudo(deps: DepsMut, env: Env, msg: SudoMsg) -> Result<Response, ContractE
             callback_gas_limit,
             mode,
             allowlist_enabled,
+            min_after,
+            max_after,
         ),
     }
 }
@@ -404,6 +446,8 @@ fn sudo_set_config(
     callback_gas_limit: Option<u64>,
     mode: Option<OperationalMode>,
     allowlist_enabled: Option<bool>,
+    min_after: Option<Timestamp>,
+    max_after: Option<Timestamp>,
 ) -> Result<Response, ContractError> {
     set_config_unchecked(
         deps,
@@ -416,6 +460,8 @@ fn sudo_set_config(
         callback_gas_limit,
         mode,
         allowlist_enabled,
+        min_after,
+        max_after,
     )
 }
 
@@ -483,6 +529,8 @@ fn set_config_unchecked(
     callback_gas_limit: Option<u64>,
     mode: Option<OperationalMode>,
     allowlist_enabled: Option<bool>,
+    min_after: Option<Timestamp>,
+    max_after: Option<Timestamp>,
 ) -> Result<Response, ContractError> {
     let config = CONFIG.load(deps.storage)?;
 
@@ -508,6 +556,15 @@ fn set_config_unchecked(
     let current_allowlist_enabled = config.allowlist_enabled.unwrap_or_default();
     let allowlist_enabled = allowlist_enabled.unwrap_or(current_allowlist_enabled);
 
+    let min_after = match min_after {
+        Some(new_value) => Some(new_value),
+        None => config.min_after,
+    };
+    let max_after = match max_after {
+        Some(new_value) => Some(new_value),
+        None => config.max_after,
+    };
+
     let new_config = Config {
         manager,
         prices,
@@ -518,6 +575,8 @@ fn set_config_unchecked(
         nois_beacon_price_updated,
         mode,
         allowlist_enabled: Some(allowlist_enabled),
+        min_after,
+        max_after,
     };
 
     CONFIG.save(deps.storage, &new_config)?;
@@ -1095,6 +1154,54 @@ mod tests {
     }
 
     #[test]
+    fn get_randomness_after_fails_for_after_values_out_of_range() {
+        let mut deps = setup(None);
+
+        // Requires a channel to forward requests to
+        setup_channel(deps.as_mut());
+
+        let instantiate_time = mock_env().block.time;
+
+        // after == instantiate_time works
+        let msg = ExecuteMsg::GetRandomnessAfter {
+            after: instantiate_time,
+            job_id: "foo".to_string(),
+        };
+        let info = mock_info("dapp", &coins(22334455, "unoisx"));
+        execute(deps.as_mut(), mock_env(), info, msg).unwrap();
+
+        // after < instantiate_time fails
+        let msg = ExecuteMsg::GetRandomnessAfter {
+            after: instantiate_time.minus_nanos(1),
+            job_id: "foo".to_string(),
+        };
+        let info = mock_info("dapp", &coins(22334455, "unoisx"));
+        let err = execute(deps.as_mut(), mock_env(), info, msg).unwrap_err();
+        match err {
+            ContractError::AfterTooLow { min_after, after } => {
+                assert_eq!(min_after, instantiate_time);
+                assert_eq!(after, instantiate_time.minus_nanos(1));
+            }
+            err => panic!("Unexpected error: {:?}", err),
+        }
+
+        // after has one 0 too much
+        let msg = ExecuteMsg::GetRandomnessAfter {
+            after: Timestamp::from_nanos(15717974198793055330),
+            job_id: "foo".to_string(),
+        };
+        let info = mock_info("dapp", &coins(22334455, "unoisx"));
+        let err = execute(deps.as_mut(), mock_env(), info, msg).unwrap_err();
+        match err {
+            ContractError::AfterTooHigh { max_after, after } => {
+                assert_eq!(max_after, Timestamp::from_nanos(1887157419879305533));
+                assert_eq!(after, Timestamp::from_nanos(15717974198793055330));
+            }
+            err => panic!("Unexpected error: {:?}", err),
+        }
+    }
+
+    #[test]
     fn set_config_works() {
         let mut deps = setup(None);
 
@@ -1114,6 +1221,8 @@ mod tests {
             callback_gas_limit: None,
             mode: None,
             allowlist_enabled: None,
+            min_after: None,
+            max_after: None,
         };
         execute(deps.as_mut(), mock_env(), mock_info(CREATOR, &[]), msg).unwrap();
         let ConfigResponse { config } =
@@ -1129,6 +1238,8 @@ mod tests {
             callback_gas_limit: None,
             mode: None,
             allowlist_enabled: Some(true),
+            min_after: None,
+            max_after: None,
         };
         execute(deps.as_mut(), mock_env(), mock_info(CREATOR, &[]), msg).unwrap();
         let ConfigResponse { config } =
@@ -1150,6 +1261,8 @@ mod tests {
             callback_gas_limit: None,
             mode: None,
             allowlist_enabled: Some(false),
+            min_after: None,
+            max_after: None,
         };
         execute(deps.as_mut(), mock_env(), mock_info(CREATOR, &[]), msg).unwrap();
         let ConfigResponse { config } =
@@ -1171,6 +1284,8 @@ mod tests {
             callback_gas_limit: Some(800_000),
             mode: None,
             allowlist_enabled: None,
+            min_after: None,
+            max_after: None,
         };
         execute(deps.as_mut(), mock_env(), mock_info(CREATOR, &[]), msg).unwrap();
         let ConfigResponse { config } =
@@ -1283,6 +1398,8 @@ mod tests {
             callback_gas_limit: None,
             mode: None,
             allowlist_enabled: Some(false),
+            min_after: None,
+            max_after: None,
         };
 
         let err = execute(deps.as_mut(), mock_env(), mock_info("dapp", &[]), msg).unwrap_err();