+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to
 ### Changed
 
 - proxy: Streamline event attributes
+- proxy: Add `min_after`/`max_after` configs in proxy to ensure the `after`
+  value for RequestBeacon packets is in a somewhat reasonable range.
 
 ## [0.13.4] - 2023-05-17